# internal/server/types.go
// Package server provides the HTTP server for the InfiniteTalk API.
// It includes handlers, middleware, routes, and DTOs separated from domain types.
package server

// CreateJobRequest is the HTTP request body for creating a new job.
type CreateJobRequest struct {
	// ImageBase64 is the base64-encoded source image.
	ImageBase64 string `json:"image_base64" validate:"required,base64"`
	// AudioBase64 is the base64-encoded source audio.
	// The audio will be processed and split into WAV PCM (pcm_s16le) chunks
	// to ensure compatibility with RunPod workers (PyAV/librosa).
	// Supported input formats: WAV, MP3, AAC, and other ffmpeg-compatible formats.
	AudioBase64 string `json:"audio_base64" validate:"required,base64"`
	// Width is the target video width.
	Width int `json:"width" validate:"required,min=1,max=4096"`
	// Height is the target video height.
	Height int `json:"height" validate:"required,min=1,max=4096"`
	// Prompt is the text prompt for video generation. Defaults to "A person talking naturally" if not provided.
	Prompt string `json:"prompt" validate:"omitempty"`
	// Provider specifies the video generation provider ("runpod" or "beam"). Defaults to "runpod".
	Provider string `json:"provider" validate:"omitempty,oneof=runpod beam"`
	// PushToS3 indicates whether to upload the final video to S3.
	PushToS3 bool `json:"push_to_s3"`
	// DryRun skips RunPod calls and completes after preprocessing.
	DryRun bool `json:"dry_run"`
	// ForceOffload forces offload on the provider. Defaults to true if not specified.
	// Use a pointer to distinguish between explicit false and not provided.
	ForceOffload *bool `json:"force_offload,omitempty"`
	// LongVideo enables V2V (Video-to-Video) mode for long-duration videos.
	// Only supported by the Beam provider.
	LongVideo bool `json:"long_video"`
}

// CreateJobResponse is the HTTP response after creating a job.
type CreateJobResponse struct {
	// ID is the unique identifier for the created job.
	ID string `json:"id"`
	// Status is the initial job status.
	Status string `json:"status"`
}

// JobResponse is the HTTP response for getting job details.
type JobResponse struct {
	// ID is the unique identifier for the job.
	ID string `json:"id"`
	// Provider is the video generation provider used for this job.
	Provider string `json:"provider"`
	// Status is the current job status.
	Status string `json:"status"`
	// Progress is the percentage of completion (0-100).
	Progress int `json:"progress"`
	// Error contains any error message if the job failed.
	Error string `json:"error,omitempty"`
	// VideoBase64 is the base64-encoded video content (if push_to_s3=false and completed).
	VideoBase64 string `json:"video_base64,omitempty"`
	// VideoURL is the S3 URL of the output video (if push_to_s3=true and completed).
	VideoURL string `json:"video_url,omitempty"`
}

// ErrorResponse is the standard error response format.
type ErrorResponse struct {
	// Error is the human-readable error message.
	Error string `json:"error"`
	// Code is the error code for programmatic handling.
	Code string `json:"code"`
}

// HealthResponse is the HTTP response for the health check endpoint.
type HealthResponse struct {
	// Status is the health status of the service.
	Status string `json:"status"`
}